Main note comp => NotesIndex and NoteCont/ImageCont their own comp
