+++ src/app/page.tsx
@@ -1,13 +1,13 @@
 "use client";
 import { Provider } from "react-redux";
 import store from "../redux/store/store";
-import NotesContainer from "./components/ui/NotesContainer";
+import NotesIndex from "./components/NotesIndex";
 
 export default function Home() {
   return (
     <main>
       <Provider store={store}>
-        <NotesContainer />
+        <NotesIndex />
       </Provider>
     </main>
   );
